feat(Chantiers): implement budget tracking and calculation system
- Add budget calculation methods for estimated and actual budgets
- Update ChantierObserver to trigger budget updates on related model changes
- Improve dashboard statistics with real budget data
- Enhance factory with realistic budget values
- Add budget progress visualization in resum view

+++ app/Livewire/Chantiers/Widgets/NbChantiersStats.php
@@ -23,7 +23,7 @@ final class NbChantiersStats extends StatsOverviewWidget
             Stat::make('Nombre total de tache active', ChantierTask::where('status', '!=', StatusChantierTask::Todo)->get()->count())
                 ->icon('data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIyNCIgaGVpZ2h0PSIyNCIgdmlld0JveD0iMCAwIDI0IDI0IiBmaWxsPSJub25lIiBzdHJva2U9ImN1cnJlbnRDb2xvciIgc3Ryb2tlLXdpZHRoPSIyIiBzdHJva2UtbGluZWNhcD0icm91bmQiIHN0cm9rZS1saW5lam9pbj0icm91bmQiIGNsYXNzPSJsdWNpZGUgbHVjaWRlLWxpc3QtdG9kby1pY29uIGx1Y2lkZS1saXN0LXRvZG8iPjxyZWN0IHg9IjMiIHk9IjUiIHdpZHRoPSI2IiBoZWlnaHQ9IjYiIHJ4PSIxIi8+PHBhdGggZD0ibTMgMTcgMiAyIDQtNCIvPjxwYXRoIGQ9Ik0xMyA2aDgiLz48cGF0aCBkPSJNMTMgMTJoOCIvPjxwYXRoIGQ9Ik0xMyAxOGg4Ii8+PC9zdmc+'),
 
-            Stat::make('Total Dépense / Budget estimé', Number::currency(Chantiers::sum('budget_estime'), 'EUR', 'fr', 2).'/'.Number::currency(ChantierDepense::sum('montant'), 'EUR', 'fr', 2))
+            Stat::make('Total Dépense / Budget estimé', Number::currency((float) Chantiers::sum('budget_estime'), 'EUR', 'fr', 2).'/'.Number::currency((float) ChantierDepense::sum('montant'), 'EUR', 'fr', 2))
                 ->icon('data:image/svg+xml;base64,PHN2ZyB4bWxucz0iaHR0cDovL3d3dy53My5vcmcvMjAwMC9zdmciIHdpZHRoPSIyNCIgaGVpZ2h0PSIyNCIgdmlld0JveD0iMCAwIDI0IDI0IiBmaWxsPSJub25lIiBzdHJva2U9ImN1cnJlbnRDb2xvciIgc3Ryb2tlLXdpZHRoPSIyIiBzdHJva2UtbGluZWNhcD0icm91bmQiIHN0cm9rZS1saW5lam9pbj0icm91bmQiIGNsYXNzPSJsdWNpZGUgbHVjaWRlLWV1cm8taWNvbiBsdWNpZGUtZXVybyI+PHBhdGggZD0iTTQgMTBoMTIiLz48cGF0aCBkPSJNNCAxNGg5Ii8+PHBhdGggZD0iTTE5IDZhNy43IDcuNyAwIDAgMC01LjItMkE3LjkgNy45IDAgMCAwIDYgMTJjMCA0LjQgMy41IDggNy44IDggMiAwIDMuOC0uOCA1LjItMiIvPjwvc3ZnPg=='),
         ];
     }
